+++ packages/expo/src/async-require/setupHMR.ts
@@ -1,10 +1,6 @@
 import HMRClient from './hmr';
 
-if (
-  typeof window !== 'undefined' &&
-  // @ts-expect-error: Added via react-native-webview
-  typeof window.$$EXPO_INITIAL_PROPS !== 'undefined'
-) {
+if (typeof window !== 'undefined' && typeof window.$$EXPO_INITIAL_PROPS !== 'undefined') {
   // Sets up developer tools for web platforms when running in a webview. This ensures that logs are visible in the terminal.
   // We assume full control over the console and send JavaScript logs to Metro.
   const LEVELS = [