+++ packages/expo/src/dom/dom-entry.tsx
@@ -7,18 +7,11 @@ import { JSONValue } from './dom.types';
 import { addEventListener, getActionsObject } from './marshal';
 import registerRootComponent from '../launch/registerRootComponent';
 
-interface MarshalledProps {
+export interface MarshalledProps {
   names: string[];
   props: Record<string, JSONValue>;
   [key: string]: undefined | JSONValue;
 }
-
-interface WindowType {
-  $$EXPO_DOM_HOST_OS?: string;
-  $$EXPO_INITIAL_PROPS?: MarshalledProps;
-}
-
-declare let window: WindowType;
 
 const ACTIONS = getActionsObject!();
 
@@ -98,11 +91,10 @@ export function registerDOMComponent(AppModule: any) {
       if (!marshalledProps.names) return {};
       // Create a named map { [name: string]: ProxyFunction }
       // TODO(@kitten): Unclear how this is typed or shaped
-      return Object.fromEntries(
-        marshalledProps.names.map((key: string) => {
-          return [key, ACTIONS[key]];
-        })
-      );
+      return marshalledProps.names.reduce((acc: Record<string, any>, key: string) => {
+        acc[key] = ACTIONS[key];
+        return acc;
+      }, {});
     }, [marshalledProps.names]);
 
     return <AppModule {...props} {...(marshalledProps.props || {})} {...proxyActions} />;